requirements.txt
click
openstacksdk
prometheus_client
pyyaml
